Fix cyclic dep
