+++ components/director/internal/repo/exist.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/kyma-incubator/compass/components/director/internal/domain/tenant"
+	tnt "github.com/kyma-incubator/compass/components/director/pkg/tenant"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/log"
 
@@ -49,10 +49,10 @@ func (g *universalExistQuerier) Exists(ctx context.Context, tenantID string, con
 		return false, apperrors.NewTenantRequiredError()
 	}
 
-	switch tenant.LoadIsolationTypeFromContext(ctx) {
-	case tenant.RecursiveIsolationType:
+	switch tnt.LoadIsolationTypeFromContext(ctx) {
+	case tnt.RecursiveIsolationType:
 		conditions = append(Conditions{NewTenantIsolationCondition(*g.tenantColumn, tenantID)}, conditions...)
-	case tenant.SimpleIsolationType:
+	case tnt.SimpleIsolationType:
 		conditions = append(Conditions{NewEqualCondition(*g.tenantColumn, tenantID)}, conditions...)
 	}
